ai-edu-ai-service/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# LLM
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
dashscope>=1.14.0  # 通义千问SDK (百炼)
zhipuai>=2.0.0     # 智谱SDK
openai>=1.0.0      # DeepSeek 兼容 OpenAI 协议

# Vector Database
pymilvus>=2.3.0

# Message Queue
pika>=1.3.2

# Utils
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Knowledge Graph
neo4j>=5.0.0
rdflib>=7.0.0
jieba>=0.42.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.28.0